+++ changelog.md
@@ -6,6 +6,7 @@
   - Code Editor:
     - Code complete for ShadowMUD inheritable files, simple begin to type inherit and it will suggest a list of predefined inherits
     - Drag and drop reorder opened files
+  - Backup: add debug output when doing client save
 - **Fixed:**
   - Code Editor:
     - Editor did not correctly resize when output window resized